requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
google-auth>=2.16.0
google-cloud-bigquery>=3.11.0
authlib>=1.3.0
requests>=2.31.0
